create revoke scopes action
