+++ src/app/services/actions/permissions-action-creator.spec.ts
@@ -18,10 +18,8 @@ import { ACCOUNT_TYPE } from '../graph-constants';
 import { consentToScopes } from '../slices/auth.slice';
 import { fetchScopes } from '../slices/scopes.slice';
 import { mockThunkMiddleware } from './mockThunkMiddleware';
-import {
-  revokeScopes
-} from './permissions-action-creator';
 import { RevokePermissionsUtil } from './permissions-action-creator.util';
+import { revokeScopes } from './revoke-scopes.action';
 
 let mockStore = configureMockStore([mockThunkMiddleware]);
