+++ src/app/services/actions/permissions-action-creator.ts
@@ -1,27 +1,15 @@
 
-import { componentNames, eventTypes, telemetry } from '../../../telemetry';
 import { AppAction } from '../../../types/action';
 import { IOAuthGrantPayload, IPermissionGrant } from '../../../types/permissions';
 import { IUser } from '../../../types/profile';
-import { RevokeScopesError } from '../../utils/error-utils/RevokeScopesError';
 import { translateMessage } from '../../utils/translate-messages';
-import {
-  DEFAULT_USER_SCOPES,
-  REVOKING_PERMISSIONS_REQUIRED_SCOPES
-} from '../graph-constants';
 import {
   GET_ALL_PRINCIPAL_GRANTS_ERROR,
   GET_ALL_PRINCIPAL_GRANTS_PENDING,
-  GET_ALL_PRINCIPAL_GRANTS_SUCCESS,
-  REVOKE_SCOPES_ERROR,
-  REVOKE_SCOPES_PENDING,
-  REVOKE_SCOPES_SUCCESS
+  GET_ALL_PRINCIPAL_GRANTS_SUCCESS
 } from '../redux-constants';
-import {
-  getConsentedScopesSuccess
-} from '../slices/auth.slice';
 import { setQueryResponseStatus } from '../slices/query-status.slice';
-import { REVOKE_STATUS, RevokePermissionsUtil } from './permissions-action-creator.util';
+import { RevokePermissionsUtil } from './permissions-action-creator.util';
 
 export function getAllPrincipalGrantsPending(response: boolean) {
   return {
@@ -45,153 +33,6 @@ export function getAllPrincipalGrantsError(response: object): AppAction {
 
 }
 
-export function revokeScopesPending(): AppAction {
-  return {
-    type: REVOKE_SCOPES_PENDING,
-    response: null
-  }
-}
-
-export function revokeScopesSuccess(): AppAction {
-  return {
-    type: REVOKE_SCOPES_SUCCESS,
-    response: null
-  }
-}
-
-export function revokeScopesError(): AppAction {
-  return {
-    type: REVOKE_SCOPES_ERROR,
-    response: null
-  }
-}
-
-interface IPermissionUpdate {
-  permissionBeingRevokedIsAllPrincipal: boolean;
-  userIsTenantAdmin: boolean;
-  revokePermissionUtil: RevokePermissionsUtil;
-  grantsPayload: IOAuthGrantPayload;
-  profile: IUser;
-  permissionToRevoke: string;
-  newScopesArray: string[];
-  retryCount: number;
-  retryDelay: number;
-  dispatch: Function;
-}
-export function revokeScopes(permissionToRevoke: string) {
-  return async (dispatch: Function, getState: Function) => {
-    const { auth: { consentedScopes }, profile } = getState();
-    const requiredPermissions = REVOKING_PERMISSIONS_REQUIRED_SCOPES.split(' ');
-    const defaultUserScopes = DEFAULT_USER_SCOPES.split(' ');
-    dispatch(revokeScopesPending());
-    dispatchScopesStatus(dispatch, 'Please wait while we revoke this permission', 'Revoking ', 0);
-    const revokePermissionUtil = await RevokePermissionsUtil.initialize(profile.id);
-
-    if (!consentedScopes || consentedScopes.length === 0) {
-      dispatch(revokeScopesError());
-      trackRevokeConsentEvent(REVOKE_STATUS.preliminaryChecksFail, permissionToRevoke);
-      return;
-    }
-
-    const newScopesArray: string[] = consentedScopes.filter((scope: string) => scope !== permissionToRevoke);
-
-    try {
-      const { userIsTenantAdmin, permissionBeingRevokedIsAllPrincipal, grantsPayload } = await revokePermissionUtil.
-        getUserPermissionChecks({ consentedScopes, requiredPermissions, defaultUserScopes, permissionToRevoke });
-      const retryCount = 0;
-      const retryDelay = 100;
-      const permissionsUpdateObject: IPermissionUpdate = {
-        permissionBeingRevokedIsAllPrincipal, userIsTenantAdmin, revokePermissionUtil, grantsPayload,
-        profile, permissionToRevoke, newScopesArray, retryCount, dispatch, retryDelay
-      }
-
-      const updatedScopes = await updatePermissions(permissionsUpdateObject)
-
-      if (updatedScopes) {
-        dispatchScopesStatus(dispatch, 'Permission revoked', 'Success', 4);
-        dispatch(getConsentedScopesSuccess(updatedScopes));
-        dispatch(revokeScopesSuccess());
-        trackRevokeConsentEvent(REVOKE_STATUS.success, permissionToRevoke);
-      }
-      else {
-        throw new RevokeScopesError({
-          errorText: 'Scopes not updated', statusText: 'An error occurred when unconsenting',
-          status: '500', messageType: 1
-        })
-      }
-    }
-    catch (errorMessage: any) {
-      if (errorMessage instanceof RevokeScopesError || errorMessage instanceof Function) {
-        const { errorText, statusText, status, messageType } = errorMessage
-        dispatchScopesStatus(dispatch, statusText, status, messageType);
-        const permissionObject = {
-          permissionToRevoke,
-          statusCode: statusText,
-          status: errorText
-        }
-        trackRevokeConsentEvent(REVOKE_STATUS.failure, permissionObject);
-      }
-      else {
-        const { code, message } = errorMessage;
-        trackRevokeConsentEvent(REVOKE_STATUS.failure, 'Failed to revoke consent');
-        dispatchScopesStatus(dispatch, message ? message : 'Failed to revoke consent', code ? code : 'Failed', 1);
-      }
-    }
-  }
-}
-
-async function updatePermissions(permissionsUpdateObject: IPermissionUpdate): Promise<string[] | null> {
-  const {
-    permissionBeingRevokedIsAllPrincipal, userIsTenantAdmin, revokePermissionUtil, grantsPayload,
-    profile, permissionToRevoke, newScopesArray, retryCount, dispatch, retryDelay } = permissionsUpdateObject;
-  let isRevokeSuccessful;
-  const maxRetryCount = 7;
-  const newScopesString = newScopesArray.join(' ');
-
-  if (permissionBeingRevokedIsAllPrincipal && userIsTenantAdmin) {
-    isRevokeSuccessful = await revokePermissionUtil.getUpdatedAllPrincipalPermissionGrant(grantsPayload,
-      permissionToRevoke);
-  } else {
-    isRevokeSuccessful = await revokePermissionUtil.updateSinglePrincipalPermissionGrant(grantsPayload, profile,
-      newScopesString);
-  }
-
-  if (isRevokeSuccessful) {
-    return newScopesString.split(' ');
-  }
-  else if ((retryCount < maxRetryCount) && !isRevokeSuccessful) {
-    await new Promise(resolve => setTimeout(resolve, retryDelay * 2));
-    dispatchScopesStatus(dispatch, 'We are retrying the revoking operation', 'Retrying', 5);
-
-    permissionsUpdateObject.retryCount += 1;
-    return updatePermissions(permissionsUpdateObject);
-  }
-  else {
-    return null;
-  }
-
-}
-
-const dispatchScopesStatus = (dispatch: Function, statusText: string, status: string, messageType: number) => {
-  dispatch(revokeScopesError());
-  dispatch(
-    setQueryResponseStatus({
-      statusText: translateMessage(status),
-      status: translateMessage(statusText),
-      ok: false,
-      messageType
-    })
-  )
-}
-
-const trackRevokeConsentEvent = (status: string, permissionObject: any) => {
-  telemetry.trackEvent(eventTypes.BUTTON_CLICK_EVENT, {
-    componentName: componentNames.REVOKE_PERMISSION_CONSENT_BUTTON,
-    permissionObject,
-    status
-  });
-}
-
 export function fetchAllPrincipalGrants() {
   return async (dispatch: Function, getState: Function) => {
     dispatch(getAllPrincipalGrantsPending(true));
@@ -204,12 +45,17 @@ export function fetchAllPrincipalGrants() {
         dispatch(getAllPrincipalGrantsPending(true));
         const requestCounter = 0;
 
-        await checkScopesConsentType(servicePrincipalAppId, tenantWideGrant, revokePermissionUtil,
-          consentedScopes, profile, requestCounter, dispatch);
+        await checkScopesConsentType(servicePrincipalAppId, tenantWideGrant,
+          revokePermissionUtil, consentedScopes, profile, requestCounter, dispatch);
       }
       else {
         dispatch(getAllPrincipalGrantsPending(false));
-        dispatchScopesStatus(dispatch, 'Permissions', 'You require the following permissions to read', 0)
+        setQueryResponseStatus({
+          statusText: translateMessage('Permissions'),
+          status: translateMessage('You require the following permissions to read'),
+          ok: false,
+          messageType: 0
+        })
       }
     } catch (error: any) {
       dispatch(getAllPrincipalGrantsPending(false));
